Add files via upload
I have done the Accenture internship in virtual for data analyst. as per the KIP'S need to clean and format the data to bring meaningful  data and need to find the top performing categories using some calculation and pivot table in Excel to client understand were the more profit they were getting and same using the power bi i have done the visual part to easy to understand.